requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
twilio==8.10.0
python-dotenv==1.0.0